#135 - TV 및 Movie 장르 관련
- 같은 파일로 관리 되어있던 코드 다른 파일로 분리
