+++ IMAD_Project/Class/TVWorkGenre.swift
@@ -1,0 +1,22 @@
+//
+//  TVWorkGenre.swift
+//  IMAD_Project
+//
+//  Created by 유영웅 on 10/28/24.
+//
+
+import Foundation
+
+class TVWorkGenre:WorkGenre{
+    var genreType: WorkGenreType = .tv
+    var tvGenre:RecommendTVResponse
+    required init(tvGenre:RecommendTVResponse) {
+        self.tvGenre = tvGenre
+    }
+    var id:Int{ tvGenre.id }
+    var name:String?{ tvGenre.name }
+    var title:String?
+    var genreId:[Int]?{ tvGenre.genreIds }
+    var posterPath:String?{ tvGenre.posterPath }
+    var backdropPath:String?{ tvGenre.backdropPath }
+}
